Move ListIndicator + PageLayout to common folder
