+++ client/packages/requisitions/src/common/ListIndicators.tsx
@@ -4,7 +4,7 @@ import {
   RouteBuilder,
   useNavigate,
 } from '@openmsupply-client/common';
-import { IndicatorLineRowFragment } from '../../../RequestRequisition/api';
+import { IndicatorLineRowFragment } from '../RequestRequisition/api';
 
 interface ListIndicatorLineProps {
   currentIndicatorLineId?: string | null;
