+++ client/packages/requisitions/src/common/index.tsx
@@ -1,0 +1,2 @@
+export * from './ListIndicators';
+export * from './PageLayout';
